+++ src/de/omegazirkel/risingworld/tools/I18n.java
@@ -8,11 +8,12 @@ import java.io.InputStreamReader;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
+import java.util.concurrent.ConcurrentHashMap;
 
 import net.risingworld.api.Plugin;
 
 public class I18n {
-    private static I18n instance = null;
+    private static ConcurrentHashMap<String, I18n> instanceMap = new ConcurrentHashMap<String, I18n>();
     private Map<String, Properties> language = new HashMap<String, Properties>();
     private static final String defaultLanguage = "en";
 
@@ -25,19 +26,23 @@ public class I18n {
     private I18n() {
     }
 
-    public static I18n getInstance() {
-        if (instance == null) {
-            instance = new I18n();
+    public static I18n getInstanceMap() {
+        logger().warn("getInstance called, response with default instance");
+        return getInstance("default");
+    }
+
+    public static I18n getInstance(String pluginName) {
+        if (!instanceMap.containsKey(pluginName)) {
+            instanceMap.put(pluginName, new I18n());
         }
-        return instance;
+        return instanceMap.get(pluginName);
     }
 
     public static I18n getInstance(Plugin plugin) {
         logger().setLevel(s.logLevel);
-        if (instance == null) {
-            instance = new I18n();
-            instance.loadLanguageData(plugin.getPath());
-        }
+
+        I18n instance = getInstance(plugin.getDescription("name"));
+        instance.loadLanguageData(plugin.getPath());
         return instance;
     }
 
@@ -52,14 +57,14 @@ public class I18n {
         FileInputStream in;
         try {
             logger().debug("Files found: " + listOfFiles.length);
-            for (int i = 0; i < listOfFiles.length; i++) {
-                logger().debug("loading: " + listOfFiles[i].getName());
-                if (listOfFiles[i].isFile() && listOfFiles[i].getName().endsWith("properties")) {
-                    String lang = listOfFiles[i].getName().substring(0, 2);
+            for (File f : listOfFiles) {
+                logger().debug("loading: " + f.getAbsolutePath());
+                if (f.isFile() && f.getName().endsWith("properties")) {
+                    String lang = f.getName().substring(0, 2);
                     // log.out("lang: "+lang);
                     Properties lngProperties = new Properties();
                     try {
-                        in = new FileInputStream(listOfFiles[i]);
+                        in = new FileInputStream(f);
                         lngProperties.load(new InputStreamReader(in, "UTF8"));
                         in.close();
                         this.language.put(lang.toLowerCase(), lngProperties);
@@ -109,6 +114,10 @@ public class I18n {
      */
     public String get(String key, String lang) {
         try {
+            if (!this.language.containsKey(defaultLanguage)) {
+                logger().error("no default language loaded. Failed to lookup " + key);
+                return key;
+            }
             Properties lngDefaultProperties = this.language.get(defaultLanguage);
             Properties lngProperties = null;
             if (!this.language.containsKey(lang.toLowerCase())) {